* common-src/security-util.c: Fix isdigit use.
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@1837 a8d146d6-cc15-0410-8900-af154a0219e0

+++ common-src/security-util.c
@@ -2581,7 +2581,7 @@ find_port_for_service(
     int        all_numeric = 1;
 
     for (s=service; *s != '\0'; s++) {
-	if (!isdigit(*s)) {
+	if (!isdigit((int)*s)) {
 	    all_numeric = 0;
 	}
     }
